Add SPMC ChaseLev Work-stealing deque

+++ nstl/base/base_typedefs.hpp
@@ -16,7 +16,7 @@ typedef uint16_t U16;
 typedef uint32_t U32;
 typedef uint64_t U64;
 
-typedef unsigned long uintptr;
+typedef uintptr_t uintptr;
 
 typedef float F32;
 typedef double F64;
